renderer/requirements.txt
pyyaml
redis
Pillow
pyglet
zerorpc